Autocall attributes recursively
Attribute paths supplied on the command line undergoes auto calling, not just
the top level. This isn't a language feature, and isn't exposed as a builtin (as
far as I can tell), so we need to implement the behavior ourselves.

This also cleans up the nix code quite a bit.