Update common nixops options to include options passed along to nix
See 'man nixops' > "COMMON OPTIONS PASSED ALONG TO NIX".

I could not reuse them from the 'nix_common_nixos_rebuild' variable,
because some options are missing, and some options are missing the
short flag alias.

I double-checked that every option in this list is in fact a valid
option for each of the main commands.